Bundle fluent-plugin-concat plugin into the package to deal with Docker log split issue.
Docker 1.13 changed the way how it handle logs: entries longer than 16k bytes will be split to a number of entries with size <=16k. This causes a number of problems for our customers using Stackdriver like:
- fragmented log lines
- broken jsons

This aims at bundling that plugin into our package to make it easier for customer who is impacted to concatenate the logs.